Add products, products variants, product variant items to vendor section with role protections

+++ app/Http/Controllers/Backend/VendorProductImageGalleryController.php
@@ -2,24 +2,30 @@
 
 namespace App\Http\Controllers\Backend;
 
-use App\DataTables\ProductImageGalleryDataTable;
+use App\DataTables\VendorProductImageGalleryDataTable;
 use App\Http\Controllers\Controller;
 use App\Models\Product;
 use App\Models\ProductImageGallery;
 use App\Traits\ImageUploadTrait;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
-class ProductImageGalleryController extends Controller
+class VendorProductImageGalleryController extends Controller
 {
+
     use ImageUploadTrait;
     /**
      * Display a listing of the resource.
      */
-    public function showTable(String $id, ProductImageGalleryDataTable $dataTable)
+    public function showTable(String $id, VendorProductImageGalleryDataTable $dataTable)
     {
         $product = Product::find($id);
+        //Check if editor is product owner
+        if(Auth::user()->vendor->id != $product->vendor_id){
+            abort(404);
+        }
         return $dataTable->with('productId', $id)
-            ->render('admin.product.image-gallery.index', compact('product'));
+            ->render('vendor.product.image-gallery.index', compact('product'));
     }
 
     /**
@@ -35,9 +41,14 @@ class ProductImageGalleryController extends Controller
      */
     public function store(Request $request)
     {
+        $product = Product::find($request->product_id);
+        //Check if editor is product owner
+        if(Auth::user()->vendor->id != $product->vendor_id){
+            abort(404);
+        }
         $request->validate([
-           'image' => 'required',
-           'image.*' => ['image', 'max:3000']
+            'image' => 'required',
+            'image.*' => ['image', 'max:3000']
         ]);
 
         //handle multi image upload
@@ -46,7 +57,7 @@ class ProductImageGalleryController extends Controller
         foreach ($imagePaths as $imagePath){
             $productImageGallery = new ProductImageGallery();
             $productImageGallery->image = $imagePath;
-            $productImageGallery->product_id = $request->product;
+            $productImageGallery->product_id = $request->product_id;
             $productImageGallery->save();
         }
         toastr()->success('Product Image Uploaded Successfully!');
@@ -84,11 +95,13 @@ class ProductImageGalleryController extends Controller
     public function destroy(string $id)
     {
         $productImage = ProductImageGallery::findOrFail($id);
+        //Check if editor is product owner
+        if(Auth::user()->vendor->id != $productImage->product->vendor_id){
+            abort(404);
+        }
         $this->deleteImage($productImage->image);
         $productImage->delete();
 
         return response(['status' => 'success', 'message' => 'Product Image Has Been Deleted Successfully!']);
     }
-
 }
-
